better move error handling
