+++ src/ui/flutter_app/lib/services/mill/mill.dart
@@ -43,3 +43,4 @@ part 'src/recorder.dart';
 part 'src/types.dart';
 part 'src/zobrist.dart';
 part 'src/import_export_service.dart';
+part 'src/move.dart';
