+++ src/ui/flutter_app/lib/services/mill/src/types.dart
@@ -324,104 +324,6 @@ extension HistoryMoveExtension on HistoryMove {
   }
 }
 
-class Move {
-  static const _invalidMove = -1;
-
-  // Square
-  int from = 0;
-  int to = 0;
-
-  // file & rank
-  int _fromFile = 0;
-  int _fromRank = 0;
-  int _toFile = 0;
-  int _toRank = 0;
-
-  // 'move' is the UCI engine's move-string
-  final String uciMove;
-
-  // "notation" is Standard Notation
-  late final String notation;
-
-  late final _MoveType type;
-
-  // TODO: [Leptopoda] attributes should probably be made getters
-  Move(this.uciMove) {
-    if (!_isLegal) {
-      throw "Error: Invalid Move: $uciMove";
-    }
-
-    if (uciMove[0] == "-" && uciMove.length == "-(1,2)".length) {
-      // TODO: [Leptopdoa] let [_MoveType] parse the move
-      type = _MoveType.remove;
-      from = _fromFile = _fromRank = _invalidMove;
-      _toFile = int.parse(uciMove[2]);
-      _toRank = int.parse(uciMove[4]);
-      to = makeSquare(_toFile, _toRank);
-      notation = "x${_squareToWmdNotation[to]}";
-      //captured = PieceColor.none;
-    } else if (uciMove.length == "(1,2)->(3,4)".length) {
-      type = _MoveType.move;
-      _fromFile = int.parse(uciMove[1]);
-      _fromRank = int.parse(uciMove[3]);
-      from = makeSquare(_fromFile, _fromRank);
-      _toFile = int.parse(uciMove[8]);
-      _toRank = int.parse(uciMove[10]);
-      to = makeSquare(_toFile, _toRank);
-      notation = "${_squareToWmdNotation[from]}-${_squareToWmdNotation[to]}";
-    } else if (uciMove.length == "(1,2)".length) {
-      type = _MoveType.place;
-      from = _fromFile = _fromRank = _invalidMove;
-      _toFile = int.parse(uciMove[1]);
-      _toRank = int.parse(uciMove[3]);
-      to = makeSquare(_toFile, _toRank);
-      // TODO: [Leptopoda] remove stringy thing
-      notation = "${_squareToWmdNotation[to]}";
-    } else if (uciMove == "draw") {
-      assert(false, "not yet implemented"); // TODO
-      logger.v("[TODO] Computer request draw");
-    } else {
-      assert(false);
-    }
-  }
-
-  bool get _isLegal {
-    if (uciMove == "draw") {
-      return true; // TODO
-    }
-
-    if (uciMove.length > "(3,1)->(2,1)".length) return false;
-
-    const String range = "0123456789(,)->";
-
-    if (!(uciMove[0] == "(" || uciMove[0] == "-")) {
-      return false;
-    }
-
-    if (uciMove[uciMove.length - 1] != ")") {
-      return false;
-    }
-
-    for (int i = 0; i < uciMove.length; i++) {
-      if (!range.contains(uciMove[i])) return false;
-    }
-
-    if (uciMove.length == "(3,1)->(2,1)".length) {
-      if (uciMove.substring(0, 4) == uciMove.substring(7, 11)) {
-        return false;
-      }
-    }
-
-    return true;
-  }
-
-  @override
-  int get hashCode => uciMove.hashCode;
-
-  @override
-  bool operator ==(Object other) => other is Move && other.uciMove == uciMove;
-}
-
 const sqBegin = 8;
 const sqEnd = 32;
 const sqNumber = 40;
